feat(lotto): 수익률 계산 메서드 수정
정수를 나누면 결과가 정수로 나오기 때문에 총수익의 형을 double로 수정

+++ src/main/java/lotto/Calculator.java
@@ -51,8 +51,8 @@ public class Calculator {
     }
 
     public String calculateEarningRate(int purchaseAmount) {
-        int totalPrize = (5000 * fifthPrize) + (50000 * fourthPrize) + (1500000 * thirdPrize)
-                + (30000000 * secondPrize) + (2000000000 * firstPrize);
+        double totalPrize = (5000.0 * fifthPrize) + (50000.0 * fourthPrize) + (1500000.0 * thirdPrize)
+                + (30000000.0 * secondPrize) + (2000000000.0 * firstPrize);
 
         double earningRate = totalPrize / purchaseAmount * 100;
